services/admin/pyproject.toml
[project]
name = "task-planner-admin"
version = "0.6.0"
description = ""
authors = [
    {name = "chpdd",email = "chupilko_dd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.40,<2.1.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.1,<2.12.0)",
    "pydantic-settings (>=2.9.1,<2.10.0)",
    "alembic (>=1.15.2,<1.16.0)",
    "pyjwt (>=2.10.1,<2.11.0)",
    "passlib (>=1.7.4,<1.8.0)",
    "bcrypt (>=4.3.0,<4.4.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=6.2.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest-dotenv = "^0.5.2"
ruff = "^0.14.10"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
env_files = [".test.env"]
env_override_existing_values = 1
